requirements.txt

bcrypt==3.1.4
click>=8.1.1
coverage==4.4.2
eventlet==0.31.0
Flask>=1.1.2
Flask-Bcrypt==1.0.1

Flask-Migrate==2.1.1
Flask-Script==2.0.6

Flask-Testing==0.7.1


Jinja2>=2.10.1
Mako==1.0.7
psycopg2-binary==2.8.5
pycparser==2.18

python-dateutil==2.8.2
pytz==2020.1



Werkzeug>=0.15.5

marshmallow-sqlalchemy>=0.28.1
jsonschema==4.16.0
Flask-SQLAlchemy
flask_marshmallow>=0.14.0
flasgger>=0.9.5
apispec>=6.0.2
python_dotenv>=0.21.0
pymongo>=4.3.2
passlib>=1.7.4
flask_jwt_extended>=4.4.4
flask_mail>=0.9.1
flask_restful>=0.3.9
flask_restx>=1.0.3
flask-cors>=3.0.10
numpy>=1.21.5

scikit-surprise>=1.1.3
surprise>=0.1